[Julia] Use integer division in sudoku to match C behavior
The current code uses `floor(i / 3)`. With `i::Int`, this promotes to `Float64`, but then needs to be implicitly `convert`ed back to `Int` to get stored in the destination array. To better match the behavior of the C code for which `/` performs integer division rounding toward 0, we can use the integer division function `div` which avoids the intermediate promotion to `Float64` and rounds toward 0 by default.