Julia bedcov: Use more efficient bitshift
In C, a << b where b is larger than the number of bits in a is undefined
behaviour. This allows the compiler to optimise it to a single shift instruction.
In contrast, in Julia, it returns zero.
The more efficient wrapping shift can be implemented in Julia and used instead.