Minor speed improvements to sudoku with more idomatic Perl
Changing interpolating double quotes to single quotes.
Moved three-part for loop to range operator for speed.
Using print as a list operator and omitting the postfix for.
Trinary is slightly faster than equivalent if-else in assignment.
Array assignment using `x` instead of for loop is faster.
`unpack` to an array much better than always substr'ing through string.
Regex is slow and unneeded.